refactor: jwtservice -> domainservice로 변경 (#53)
