+++ src/main/java/com/aztgg/api/auth/application/AuthService.java
@@ -4,6 +4,7 @@ import com.aztgg.api.auth.application.dto.response.LoginResponse;
 import com.aztgg.api.auth.application.exception.UserException;
 import com.aztgg.api.auth.application.strategy.AuthCredentials;
 import com.aztgg.api.auth.application.strategy.AuthStrategy;
+import com.aztgg.api.auth.domain.JwtTokenDomainService;
 import com.aztgg.api.auth.domain.RefreshTokenRepository;
 import com.aztgg.api.auth.domain.User;
 import com.aztgg.api.auth.domain.UserDomainService;
@@ -27,7 +28,7 @@ import java.util.stream.Collectors;
 public class AuthService {
 
     private final UserDomainService userDomainService;
-    private final JwtService jwtService;
+    private final JwtTokenDomainService jwtTokenDomainService;
     private final RefreshTokenRepository refreshTokenRepository;
     private final CookieUtil cookieUtil;
     private final Map<Class<? extends AuthCredentials>, AuthStrategy> strategyMap;
@@ -40,12 +41,12 @@ public class AuthService {
     public AuthService(
         UserDomainService userDomainService,
         List<AuthStrategy> authStrategies,
-        JwtService jwtService,
+        JwtTokenDomainService jwtTokenDomainService,
         RefreshTokenRepository refreshTokenRepository,
         CookieUtil cookieUtil
     ) {
         this.userDomainService = userDomainService;
-        this.jwtService = jwtService;
+        this.jwtTokenDomainService = jwtTokenDomainService;
         this.refreshTokenRepository = refreshTokenRepository;
         this.cookieUtil = cookieUtil;
 
@@ -63,8 +64,8 @@ public class AuthService {
         }
 
         User user = strategy.authenticate(credentials);
-        String accessToken = jwtService.generateToken(user);
-        String refreshToken = jwtService.generateRefreshToken(user);
+        String accessToken = jwtTokenDomainService.generateAccessToken(user);
+        String refreshToken = jwtTokenDomainService.generateRefreshToken(user);
 
         refreshTokenRepository.save(user.getUsername(), refreshToken);
         setRefreshTokenCookie(refreshToken);
@@ -76,8 +77,8 @@ public class AuthService {
         String username = refreshTokenRepository.validateRefreshToken(refreshToken);
         User user = userDomainService.findUserByUsername(username);
 
-        String newAccessToken = jwtService.generateToken(user);
-        String newRefreshToken = jwtService.generateRefreshToken(user);
+        String newAccessToken = jwtTokenDomainService.generateAccessToken(user);
+        String newRefreshToken = jwtTokenDomainService.generateRefreshToken(user);
 
         refreshTokenRepository.rotateRefreshToken(refreshToken, username, newRefreshToken);
         setRefreshTokenCookie(newRefreshToken);
